+++ src/lib/isFloat.js
@@ -8,9 +8,10 @@ export default function isFloat(str, options) {
   if (str === '' || str === '.' || str === '-' || str === '+') {
     return false;
   }
+  const value = parseFloat(str.replace(',', '.'));
   return float.test(str) &&
-    (!options.hasOwnProperty('min') || str >= options.min) &&
-    (!options.hasOwnProperty('max') || str <= options.max) &&
-    (!options.hasOwnProperty('lt') || str < options.lt) &&
-    (!options.hasOwnProperty('gt') || str > options.gt);
+    (!options.hasOwnProperty('min') || value >= options.min) &&
+    (!options.hasOwnProperty('max') || value <= options.max) &&
+    (!options.hasOwnProperty('lt') || value < options.lt) &&
+    (!options.hasOwnProperty('gt') || value > options.gt);
 }