+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+#### HEAD
+
+- Allow floating points in `isNumeric()`
+  ([#810](https://github.com/chriso/validator.js/pull/810))
+- Disallow GMail addresses with multiple consecutive dots, or leading/trailing dots
+  ([#820](https://github.com/chriso/validator.js/pull/820))
+- Added an `isRFC3339()` validator
+  ([#816](https://github.com/chriso/validator.js/pull/816))
+- Reject domain parts longer than 63 octets in `isFQDN()`, `isURL()` and `isEmail()`
+  ([bb3e542](https://github.com/chriso/validator.js/commit/bb3e542))
+- Added a new Amex prefix to `isCreditCard()`
+  ([#805](https://github.com/chriso/validator.js/pull/805))
+- Fixed `isFloat()` min/max/gt/lt filters when a locale with a comma decimal is used
+  ([2b70821](https://github.com/chriso/validator.js/commit/2b70821))
+- New locales
+  ([#803](https://github.com/chriso/validator.js/pull/803))
+
 #### 9.4.1
 
 - Patched a [REDOS](https://en.wikipedia.org/wiki/ReDoS) vulnerability in `isDataURI`